+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Changes to Calva.
 - [Support for custom project/workflow commands](https://github.com/BetterThanTomorrow/calva/issues/281)
 
 ## [Unreleased]
+- [The extension should specify the default schemes for document selectors] (https://github.com/BetterThanTomorrow/calva/issues/368)
+- [Make code more robust in case Jack-in task fails.] (https://github.com/BetterThanTomorrow/calva/issues/367)
+- [Support Jack-in without file open for single-rooted workspace] (https://github.com/BetterThanTomorrow/calva/issues/366)
 - [Show argument list of fn](https://github.com/BetterThanTomorrow/calva/issues/238)
 - [Insourcing @tonsky's Clojue Warrior, now named Calva Highlight](https://github.com/BetterThanTomorrow/calva/pull/362)
 - [Update status bar when configuration changed](https://github.com/BetterThanTomorrow/calva/issues/358)